--- BlaBlaBla/BlaBlaBla/New Group/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  BlaBlaBla
-//
-//  Created by camila oliveira on 06/04/18.
-//  Copyright © 2018 Aliamcami. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
