--- BlaBlaBla/BlaBlaBla/New Group/pt-BR.lproj/LaunchScreen.strings
@@ -1,1 +1,0 @@
-
