--- BlaBlaBla/BlaBlaBla/New Group/pt-BR.lproj/Main.strings
@@ -1,1 +1,0 @@
-
