--- BlaBlaBla/BlaBlaBla/Protocols.swift
@@ -1,28 +1,0 @@
-//
-//  WordsLoader.swift
-//  BlaBlaBla
-//
-//  Created by camila oliveira on 06/04/18.
-//  Copyright © 2018 Aliamcami. All rights reserved.
-//
-
-import UIKit
-
-protocol WordListProtocol {
-    static var countries: [String] {get}
-    static var places: [String] {get}
-    static var adjectives: [String] {get}
-    static var animals: [String] {get}
-    static var objects: [String] {get}
-}
-
-//MARKK:- SENTENCE MAKER
-protocol SentenceGeneratorProtocol {
-    static func random() -> String
-    static var all : [() -> String] {get}
-    typealias wl = WordListProtocol
-}
-extension SentenceGeneratorProtocol{
-    ///Private init
-    private init(){self.init()}
-}
