+++ BlaBlaBla/BlaBlaBla/Source/Extensions.swift
@@ -5,7 +5,6 @@
 //  Created by camila oliveira on 06/04/18.
 //  Copyright © 2018 Aliamcami. All rights reserved.
 //
-
 import UIKit
 
 extension Array {
