+++ BlaBlaBla/BlaBlaBla/Source/Sentence Generator/EnSentenceGenerator.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2018 Aliamcami. All rights reserved.
 //
 
-import UIKit
+
 
 class EnSentenceGenerator: SentenceGeneratorProtocol {
     /*
