+++ BlaBlaBla/BlaBlaBla/Source/WordLists/WorldListProtocol.swift
@@ -1,0 +1,15 @@
+//
+//  WorldListProtocol.swift
+//  BlaBlaBla
+//
+//  Created by camila oliveira on 12/04/18.
+//  Copyright © 2018 Aliamcami. All rights reserved.
+//
+
+protocol WordListProtocol {
+    static var countries: [String] {get}
+    static var places: [String] {get}
+    static var adjectives: [String] {get}
+    static var animals: [String] {get}
+    static var objects: [String] {get}
+}
